requirements.txt
aiocontextvars==0.2.2
certifi==2020.6.20
chardet==3.0.4
contextvars==2.4
idna==2.10
immutables==0.14
loguru==0.5.3
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
XlsxWriter==1.3.6
